separate Stream and Summary entity definitions, repositories and resolvers
