+++ src/api/StreamResolver.ts
@@ -1,7 +1,8 @@
 import { Arg, Int, Query, Resolver } from 'type-graphql'
 import { Inject, Service } from 'typedi'
-import { OrderBy, OrderDirection, Streams, Summary } from '../entities'
-import { StreamRepository } from '../StreamRepository'
+import { OrderDirection } from '../entities/OrderDirection'
+import { OrderBy, Streams } from '../entities/Stream'
+import { StreamRepository } from '../repository/StreamRepository'
 
 @Resolver()
 @Service()
@@ -26,10 +27,5 @@ export class StreamResolver {
         @Arg("cursor", { nullable: true }) cursor?: string,
     ): Promise<Streams> {
         return this.repository.getStreams(ids, searchTerm, owner, orderBy, orderDirection, pageSize, cursor)
-    }
-
-    @Query(() => Summary)
-    async summary(): Promise<Summary> {
-        return this.repository.getSummary()
     }
 }
