+++ src/crawler/Crawler.ts
@@ -7,7 +7,7 @@ import pLimit from 'p-limit'
 import { DhtAddress, Stream, StreamCreationEvent, StreamMetadata, StreamPermission } from 'streamr-client'
 import { Inject, Service } from 'typedi'
 import { CONFIG_TOKEN, Config } from '../Config'
-import { StreamRepository } from '../StreamRepository'
+import { StreamRepository } from '../repository/StreamRepository'
 import { StreamrClientFacade } from '../StreamrClientFacade'
 import { collect, retry } from '../utils'
 import { NetworkNodeFacade } from './NetworkNodeFacade'
