+++ src/entities/OrderDirection.ts
@@ -1,0 +1,10 @@
+import { registerEnumType } from 'type-graphql'
+
+export enum OrderDirection {
+    ASC = 'ASC',
+    DESC = 'DESC'
+}
+
+registerEnumType(OrderDirection, {
+    name: 'OrderDirection'
+})
