+++ src/entities/Summary.ts
@@ -1,0 +1,10 @@
+import { Field, Float, Int, ObjectType } from 'type-graphql'
+
+/* eslint-disable indent */
+@ObjectType()
+export class Summary {
+    @Field(() => Int)
+    streamCount!: number
+    @Field(() => Float)
+    messagesPerSecond!: number
+}
