+++ test/end-to-end.test.ts
@@ -8,7 +8,7 @@ import Container from 'typedi'
 import { CONFIG_TOKEN } from '../src/Config'
 import { APIServer } from '../src/api/APIServer'
 import { Crawler } from '../src/crawler/Crawler'
-import { Stream } from '../src/entities'
+import { Stream } from '../src/entities/Stream'
 import { createDatabase, queryAPI } from '../src/utils'
 import { TEST_DATABASE_NAME, dropTestDatabaseIfExists } from './utils'
 import { range } from 'lodash'
